[refactor] Member 엔티티에서 email 필드 임시 비활성화
