+++ yechef/src/main/java/sejong/capston/yechef/domain/Member/dto/MemberDto.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 public class MemberDto {
   private Long id;
   private String nickname;
-  private String email;
+  //private String email;
 
 
 }
